Refactor token handling to utilize getTokenIndicesInClosedExpressionRanges
